refactor(server): split the 6.6k-line server.rs into a directory module
The second swollen organ from the 2026-07-02 health review. Extracts three
cohesive pieces out of `server.rs` into submodules under `src/server/`,
leaving the handler, router, gating, cast-enum, and resource plumbing in
`mod.rs`. The external surface (`KaiboHandler`/`ToolGating`/`router`/`wrapper`)
is unchanged, so every `crate::server::X` call site is untouched.

- **containment** — the read-scope boundary as its own doc-headed module:
  `resolve_root`, `contained`, `resolve_attachments`, as a split inherent
  `impl super::KaiboHandler` (a child module sees the handler's private
  fields). This makes the security-critical boundary a legible standalone
  unit instead of three methods buried in a 2000-line impl. The shared
  predicates it defers to (`containing_tree`, `containment_error`) stay in
  mod.rs. `resolve_root` widens private -> `pub(super)` (the minimal step so
  mod.rs can still call it across the boundary); nothing else widens.
- **config_resource** — `render_config_resource` and the doc structs nested
  in its body (`ConfigDoc`/`RuntimeDoc`/`SandboxDoc`/...).
- **render** — the job/batch/wait result-rendering cluster (`render_job`,
  `render_jobs_section`, `render_wait`, the batch-handle/window helpers,
  failure classification, `with_provenance`, `consult_result`), `pub(super)`
  where mod.rs calls them.

Pure behavior-preserving move: the containment methods are byte-identical to
the originals (verified against HEAD), no logic/string/const/test-assertion
changed. Tests distributed to the owning submodule; full suite green (308 lib
+ every integration suite, incl. the security teeth `tests/containment.rs` 19
and `tests/attach.rs` 10). Clippy clean (the one warning is pre-existing, in
telemetry.rs). Cross-family review (DeepSeek via kaibo `oneshot`) confirmed
the boundary is intact and the visibility changes are minimal/correct.

Internal refactor — no user-facing change, so no CHANGELOG entry.
